% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/options.R
\name{addOptions}
\alias{addOptions}
\title{Build Options Dataset}
\usage{
addOptions(stocksData, optionNames = NULL, underlyingNames, Date = NULL,
  impVol, strike, moneyness, callput, maturity)
}
\arguments{
\item{stocksData}{a \code{stocksData} object, the stocks dataset that contains underlying information.}

\item{optionNames}{a character vector, the symbols of the options.}

\item{underlyingNames}{a character vector of the same length as \code{optionNames}, which specifies the underlying of each option.}

\item{Date}{a Date vector, which specifies the dates of options data.}

\item{impVol}{the implied volatility of each option for each date.}

\item{strike, moneyness}{a numerical vector that specifies the strike or moneyness of the options.}

\item{callput}{a character vector which can either take \code{"call"} or \code{"put"}.}

\item{maturity}{time to maturity of the options.}
}
\value{
An object of class \code{optionsData}, which stores the structured data of options and stocks.
}
\description{
This function is the creator for an S3 class of structured dataset on options.
}
\details{
This function builds an options dataset by organizing raw options data and underlying stocks data 
into a structured form.
If \code{optionsNames==NULL}, the options will be automatically named as \code{"OP1"}, \code{"OP2"}, and so on. 
If \code{Date==NULL}, the options data is regarded as in reverse chronological order with the same dates as in \code{stocksData}.
}
\seealso{
\code{\link{stocks}} for building \code{stocksData} objects.
}

