% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/backtest.R
\name{backtest}
\alias{backtest}
\title{VaR Backtesting}
\usage{
backtest(object, testLen.years = 1)
}
\arguments{
\item{object}{an object of class \code{"VaR"}.}

\item{testLen.years}{the window size of backtesting in years.}
}
\value{
An object of class \code{"VaR.backtesting"}.
}
\description{
This function performs VaR backtesting and creates an S3 class object of backtesting results
}
\details{
Backtesting is performed by firstly compute the realized loss, and then calculate the rate that
VaR is exceeded across the backtesting window.
}
\seealso{
\code{\link{VaR}}
}

