% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/GBM.R
\name{GBM}
\alias{GBM}
\alias{GBM.default}
\alias{GBM.stocksData}
\title{GBM Model Fitting}
\usage{
GBM(...)

\method{GBM}{default}(drift, volatility, S0, corr = NULL, Names = NULL,
  rate = 0, divRate = 0, ...)

\method{GBM}{stocksData}(data, len = 5, lambda = NULL,
  method = c("unweighted", "exp-weighted"), ...)
}
\arguments{
\item{...}{arguments to be passed to or from methods.}

\item{drift}{a vector or matrix, which manually sets the drift of the model.}

\item{volatility}{a vector or matrix, which manually sets the volatility.}

\item{S0}{a vector or matrix, which specifies the spot price.}

\item{corr}{a list of correlation matrices.}

\item{Names}{a character vector that specifies the symbols of the stocks.}

\item{rate}{a numerical scalar or vector specifying the risk free rate.}

\item{divRate}{a scalar, vector or matrix, which specifies the dividend rates.}

\item{data}{an object of class or inherited from \code{stocksData}.}

\item{len}{the window size in years.}

\item{lambda}{the decay factor used by exponential weighting method. 
If \code{NULL}, the function will use the equivalence of \code{len}.}

\item{method}{a character string indicating calibration method to be used.
Must be \code{"unweighted"} (default) or \code{"exp-weighted"}, can be abbreviated.}
}
\value{
An object of class \code{"GBM"} which contains the data and the parameters.
}
\description{
A generic function which is the creator for an S3 class of multivariate GBM model.
}
\details{
This generic function is the mid-layer of the risk calculation system, which performs 
model calibration based on historical data and parameter specification. 
See model documentation for details.
}
\seealso{
\code{\link{stocks}}, \code{\link{addOptions}}, \code{\link{portfolio}}.
}

