% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/options.R, R/stocks.R
\name{merge.optionsData}
\alias{merge.optionsData}
\alias{merge.stocksData}
\title{Merge Stocks or Options Datasets}
\usage{
\method{merge}{optionsData}(x, y, ...)

\method{merge}{stocksData}(x, y, ...)
}
\arguments{
\item{x, y}{objects of class \code{stocksData} or \code{optionsData}.}

\item{...}{arguments to be passed to or from methods.}
}
\value{
An object of the same class as \code{x}.
}
\description{
Merge two stocks or options datasets by common dates.
}
\details{
If \code{x} is a \code{stocksData} object but \code{y} is an \code{optionsData} object, 
the options data in \code{y} will be lost in the merged result.
Also if \code{x} or \code{y} have been calibrated to a GBM model, that information will also be lost 
in the merged dataset.
}
\seealso{
\code{\link{stocks}}, \code{\link{addOptions}}
}

