% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/portfolio.R
\name{portfolio}
\alias{portfolio}
\title{Build Portfolio Dataset}
\usage{
portfolio(data, Names, weights, V0 = 1)
}
\arguments{
\item{data}{an object that contains the components of the portfolio.}

\item{Names}{a character vector indicating the components of the portfolio.}

\item{weights}{an object of class \code{pfWeight}, or a numerical vector (see Details).}

\item{V0}{the total value of the portfolio.}
}
\value{
An object of class \code{portfolioData}.
}
\description{
This function is the creator of an S3 class of structured dataset for a portfolio.
}
\details{
This function, together with \code{\link{stocks}} and \code{\link{addOptions}}, 
forms the bottom layer of the risk calculation system, which performs data structuring. 

This function subsets the data and only keeps the components of the portfolio 
(and the underlyings if components are options). 

The function support portfolio weights defined both in value and number of shares, which is supported by
\code{\link{portfolioWeight}} function. 
If \code{weights} is a numerical vector, 
it is regarded as portfolio weights in value, not in number of shares.
}
\seealso{
\code{\link{stocks}}, \code{\link{addOptions}}, \code{\link{GBM}}, \code{\link{portfolioWeight}}
}

