% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/portfolio.R
\name{portfolioWeight}
\alias{portfolioWeight}
\title{Define Portfolio Weights}
\usage{
portfolioWeight(weight, unit = c("value", "share"))
}
\arguments{
\item{weight}{a numerical vector that specifies the weights}

\item{unit}{a character string specifying how the weights are defined. 
One of \code{"value"} (default), or \code{"share"}. Also see \code{"Details"}.}
}
\value{
An object of class \code{pfWeight}.
}
\description{
This function is the creator of an S3 class that defines portfolio weights 
in terms of a proportion in dollar value or number of shares for a portfolio.
}
\details{
\code{portfolioWeight} is able to produce portfolio weights defined using different approaches: 
the most basic type of weight is determined by dividing the dollar value of a security 
by the total dollar value of the portfolio (\code{unit = "value"}). 
Another approach would be to divide the number of units of a given security by the total number of shares 
held in the portfolio (\code{unit = "share"}).
}
\seealso{
\code{\link{portfolio}}
}

