% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/backtest.R
\name{realizedLoss}
\alias{realizedLoss}
\alias{realizedLoss.optionsData}
\alias{realizedLoss.portfolioData}
\alias{realizedLoss.stocksData}
\title{Calcuate Realized Loss}
\usage{
realizedLoss(data, TimeLen, ...)

\method{realizedLoss}{portfolioData}(data, TimeLen, ...)

\method{realizedLoss}{optionsData}(data, TimeLen, V0 = NULL, ...)

\method{realizedLoss}{stocksData}(data, TimeLen, V0 = NULL, ...)
}
\arguments{
\item{data}{an object of structured dataset.}

\item{TimeLen}{the time horizon the losses are defined over, in days.}

\item{...}{arguments to be passed to or from methods.}

\item{V0}{the total value the losses are defined upon.}
}
\value{
A matrix of realized losses.
}
\description{
This generic function is an internal function used by \code{\link{backtest}} to calculate realized loss
given various types of dataset.
}
\seealso{
\code{\link{backtest}}
}

