% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/riskcalc-package.R
\docType{package}
\name{riskcalc-package}
\alias{riskcalc}
\alias{riskcalc-package}
\title{riskcalc: Build risk calculators for individualized prediction}
\description{
\if{html}{\figure{logo.png}{options: style='float: right' alt='logo' width='120'}}

Contains a toolkit for building Shiny-based risk calculators that can be used for individualized risk estimation and decision making. Motivated by the functionality of applications at https://riskcalc.org/.
}
\seealso{
Useful links:
\itemize{
  \item \url{https://clevelandclinicqhs.github.io/riskcalc/}
}

}
\author{
\strong{Maintainer}: Alex Zajichek \email{zajicha2@ccf.org} (\href{https://orcid.org/0000-0001-9591-6692}{ORCID})

Other contributors:
\itemize{
  \item Cleveland Clinic Quantitative Health Sciences [copyright holder, funder]
}

}
\keyword{internal}
