% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/riskcalc_fun.R
\name{riskcalc_fun}
\alias{riskcalc_fun}
\title{Create a shiny app}
\usage{
riskcalc_fun(x)
}
\arguments{
\item{x}{\ifelse{html}{\href{https://lifecycle.r-lib.org/articles/stages.html#questioning}{\figure{lifecycle-questioning.svg}{options: alt='[Questioning]'}}}{\strong{[Questioning]}} Unused argument}
}
\value{
A shiny app
}
\description{
\ifelse{html}{\href{https://lifecycle.r-lib.org/articles/stages.html#experimental}{\figure{lifecycle-experimental.svg}{options: alt='[Experimental]'}}}{\strong{[Experimental]}}
}
\examples{
if(interactive()) {
  riskcalc_fun()
}
}
