% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/stocks.R
\name{stocks}
\alias{stocks}
\title{Build Stocks Dataset}
\usage{
stocks(data = NULL, Names = NULL, date.col = 1, price.col = -1,
  rate.col = NULL, divRate.col = NULL)
}
\arguments{
\item{data}{a data frame, the raw data that contains stock prices.}

\item{Names}{a character vector, the symbols of the stocks.}

\item{date.col}{an integer or a character string that specifies the date column in data.}

\item{price.col}{a vector of integers or character strings that specifies the price columns in data.}

\item{rate.col}{an integer or a string that specifies the column for risk free rate.}

\item{divRate.col}{a vector of integers or strings that specifies the columns for dividend rates.}
}
\value{
An object of class \code{stocksData}, which stores the structured data of stocks.
}
\description{
This function is the creator for an S3 class of structured dataset on stocks.
}
\details{
This function builds a stocks dataset by organizing raw data into a structured form.
If \code{Names==NULL}, the stocks will be automatically named as \code{"S1"}, \code{"S2"}, and so on. 
If the raw data does not have a date column, \code{date.col} should be set to \code{NULL}, 
and the data is regarded as in reverse chronological order.
}

