% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/GBM.R, R/options.R, R/stocks.R
\name{subset.GBM}
\alias{subset.GBM}
\alias{subset.optionsData}
\alias{subset.stocksData}
\title{Subsetting Options or Stocks Dataset}
\usage{
\method{subset}{GBM}(x, select, ...)

\method{subset}{optionsData}(x, select, drop = TRUE, ...)

\method{subset}{stocksData}(x, select, ...)
}
\arguments{
\item{x}{object to be subsetted, of class \code{stocksData}, \code{optionsData} or \code{GBM}.}

\item{select}{a character vector specifying selected symbols of stocks and/or options.}

\item{...}{arguments to be passed to or from methods.}

\item{drop}{logical. If \code{TRUE}, the unused underlying stocks in the options dataset will be dropped.}
}
\value{
An object of the same class as \code{x}.
}
\description{
Return subset of the dataset with selected symbols.
}
\details{
If you subset an \code{optionsData} object and do not select any options, 
the result will still be an \code{optionsData} object with no options. \code{GBM} objects also can be
subsetted and parameters for selected instruments will be kept intact.
}
\seealso{
\code{\link{stocks}}, \code{\link{addOptions}}, \code{\link{GBM}}
}

