% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/VaR.R
\name{VaR}
\alias{VaR}
\alias{VaR.optionsData}
\alias{VaR.portfolioData}
\alias{VaR.stocksData}
\title{Calculate VaR with Various Methods}
\usage{
VaR(data, ...)

\method{VaR}{portfolioData}(data, TimeLen = 5, conf.level = 0.99,
  method = c("parametric", "historical", "Monte-Carlo"), lookback.years = 5,
  iteration = 10000, ...)

\method{VaR}{optionsData}(data, TimeLen = 5, conf.level = 0.99,
  method = c("parametric", "historical", "Monte-Carlo"), V0 = 10000,
  lookback.years = 5, iteration = 10000, ...)

\method{VaR}{stocksData}(data, TimeLen = 5, conf.level = 0.99,
  method = c("parametric", "historical", "Monte-Carlo"), V0 = 10000,
  lookback.years = 5, iteration = 10000, ...)
}
\arguments{
\item{data}{an object that contains structured data of stocks, options, or a portfolio.}

\item{...}{arguments to be passed to or from methods.}

\item{TimeLen}{the time horizon in days that the VaR is defined over.}

\item{conf.level}{the confidence level of VaR}

\item{method}{the method used to calculate VaR. Must be one of \code{"parametric"} (default), 
\code{"historical"}, or \code{"Monte-Carlo"}, can be abbreviated.}

\item{lookback.years}{the time horizon in years that historical simulation is based on.}

\item{iteration}{the number of iteration Monte-Carlo simulation to be performed.}

\item{V0}{the total value which VaR is measured upon.}
}
\value{
An object of class \code{"VaR"}.
}
\description{
This generic function computes VaR and creates an S3 class object that contains data and VaR results.
}
\details{
This function is able to compute portfolio VaR for portfolio data, and is also able to 
compute VaR for each individual stocks and/or options given stocks or options data. 

In parametric method and Monte-Carlo method, the portfolio components should be firstly calibrated
to a GBM model. In parametric method, the portfolio is assumed normally distributed and the options are
linearly approximated by underlying stocks. In historical and Monte-Carlo method, 
the option values are fully evaluated.

See model documentation for details about VaR calculation methods.
}
\seealso{
\code{\link{backtest}} to perform VaR backtesting.
}

