% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/risk_calculator.R
\name{risk_calculator}
\alias{risk_calculator}
\title{Build a risk calculator}
\usage{
risk_calculator(inputs, outputs, title = "", intercept = 0)
}
\arguments{
\item{inputs}{Inputs for risk calculator}

\item{outputs}{Outputs for risk calculator}

\item{title}{Title for risk calculator}

\item{intercept}{Intercept term to add to the linear combination}
}
\value{
A shiny app
}
\description{
Construct a shiny-based risk calculator given information about the inputs and outputs
}
\examples{
# Example from https://riskcalc.org/bladderCancer/
# Source code: https://github.com/ClevelandClinicQHS/riskcalc-website/tree/main/bladderCancer

# Build set of inputs
inputs <-

  # Each element is one input (risk factor) in that order
  list(
    Gender =
      list(
        label = "Gender",
        levels = c("Female", "Male"), # Indicates categorical
        weights = c(0, 0.034072352)
      ),
    AgeYr =
      list(
        label = "Age (Years)",
        range = c(20, 100), # Indicates numeric
        x_trans =
          list(
            \(x)x,
            \(x)max(x - 44, 0)**3,
            \(x)max(x - 62, 0)**3,
            \(x)max(x - 77, 0)**3
           ), # Each of these contributes to weighted sum
        weights = c(0.0041362795, 7.4691198e-06, 1.6432063e-05, 8.9629437e-06)
      ),
    Stage97 =
      list(
        label = "RC Tumor Pathology",
        levels = c("pT0", "pTa", "pTis", "pT1", "pT2", "pT3", "pT4"),
        weights = c(0, 0.87060903, 0.55456587, 1.0188265, 1.2332424, 1.8741757, 2.4566641)
      ),
    Histology =
      list(
        label = "RC Tumor Histology",
        levels = c("ADENOCARCINOMA", "SCC", "TCC"),
        weights = c(0, 0.34940226, 0.72706351)
      ),
    Grade =
      list(
        label = "RC Tumor Grade",
        levels = c("GX", "High", "Low"),
        weights = c(0, 0.17941245, 0.69780658)
      ),
    NodeResult =
      list(
        label = "Lymph Node Status",
        levels = c("NX", "Negative", "Positive"),
        weights = c(0, 0.29095514, 0.87448703)
      ),
    DxToRC =
      list(
        label = "Days Between Dx and RC (Days)",
        range = c(0, 100000),
        x_trans =
          list(
            \(x)x,
            \(x)max(x - 0.45996875, 0)**3,
            \(x)max(x - 3.5483304, 0)**3,
            \(x)max(x - 27.808397, 0)**3
           ),
        weights = c(0, 2.9994932e-05, 3.3813355e-05, 3.8184231e-06)
      )
  )

# Build outputs to display in the table
outputs <-
  list(
    res =
      list(
        label = "Percentage of 5-Year Recurrence-Free Survival",
        lp_trans = \(x) 100 * 0.7524126 ** exp(x) # Transformation for linear predictor
      )
  )

# Create the risk calculator
if(interactive()) {
  risk_calculator(
    inputs = inputs,
    outputs = outputs,
    title =
      paste(
        "Predicting 5-Year Recurrence-Free Survival",
        "after Radical Cystectomy for Bladder Cancer"
       ),
    intercept = -2.1264314
  )
}

}
