% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/risk_calculator.R
\name{risk_calculator}
\alias{risk_calculator}
\alias{risk_calculator.default}
\alias{risk_calculator.glm}
\alias{risk_calculator.coxph}
\title{Build a risk calculator}
\usage{
risk_calculator(...)

\method{risk_calculator}{default}(app_name = NULL, app_directory = NULL, ...)

\method{risk_calculator}{glm}(
  model,
  app_name = NULL,
  app_directory = NULL,
  title = "",
  citation = "",
  label = "Predicted Value",
  label_header = "Result",
  value_header = "Value",
  format = NULL,
  labels = NULL,
  levels = NULL,
  placeholders = NULL,
  ...
)

\method{risk_calculator}{coxph}(
  model,
  time,
  app_name = NULL,
  app_directory = NULL,
  title = "",
  citation = "",
  label = "Survival Probability",
  label_header = "Result",
  value_header = "Value",
  format = NULL,
  labels = NULL,
  levels = NULL,
  placeholders = NULL,
  ...
)
}
\arguments{
\item{...}{Generic arguments}

\item{app_name}{Application name (e.g., \code{"AppExample"} for www.riskcalc.org/AppExample. If no name is provided, a generic name "riskcalc_app" is used.}

\item{app_directory}{Location to export the application. If none provided, defaults to the working directory.}

\item{model}{A \code{\link[stats]{glm}} or \code{\link[survival]{coxph}} object}

\item{title}{Title for risk calculator}

\item{citation}{Citation(s) and author information}

\item{label}{Label for the calculated value}

\item{label_header}{Header label for the result descriptor}

\item{value_header}{Header label for the result value}

\item{format}{Function to format the predicted value for display}

\item{labels}{Named character vector specifying labels for any inputs}

\item{levels}{Named list of named character vectors specifying labels for any factor levels}

\item{placeholders}{Named character vector specifying range limits for numeric input variables (of the form \code{c(var="<min>-<max>")}). Invokes calls to \code{\link[shiny]{validate}} enforcing the limits.}

\item{time}{Single time point to calculate survival probability (see \code{\link[survival]{summary.survfit}})}
}
\value{
A \code{\link[shiny]{shiny}} application
}
\description{
Create a (templated) risk calculator for www.riskcalc.org, optionally populating using an existing model object
}
\examples{
# Make a data set
dat <- survival::gbsg
dat$meno <- factor(dat$meno)

# Build a model
mod <-
  survival::coxph(
    formula = survival::Surv(rfstime, status) ~ age + meno,
    data = dat
  )

# Time point of interest
time <- median(dat$rfstime) # 1084

# Create the risk calculator
if(interactive()) {
  risk_calculator(
    model = mod,
    time = time,
    title = "Risk of Death or Recurrence in Breast Cancer",
    citation =
      paste0(
        "Patrick Royston and Douglas Altman, External validation of a Cox prognostic",
        "model: principles and methods. BMC Medical Research Methodology 2013, 13:33"
      ),
    label_header = "Probability Of Death or Recurrence",
    label = paste0("At time = ", time),
    labels = c(age = "Age (years)", meno = "Menopausal Status"),
    levels = list(meno = c(`0` = "Premenopausal", `1` = "Postmenopausal")),
    placeholders = c(age = "21-80"),
    format = function(x) paste0(round(100 * (1 - x), 2), "\%")
  )
}
}
