% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/risk_calculator.R
\name{risk_calculator}
\alias{risk_calculator}
\alias{risk_calculator.glm}
\title{Build a risk calculator}
\usage{
risk_calculator(model, ...)

\method{risk_calculator}{glm}(
  model,
  title = "",
  citation = "",
  label = "Predicted Value",
  label_header = "Result",
  value_header = "Value",
  format = NULL,
  app_name = NULL,
  labels = NULL,
  levels = NULL,
  placeholders = NULL,
  ...
)
}
\arguments{
\item{model}{A \code{\link[stats]{glm}} object whose \code{\link[stats]{formula}} only has non-negated factors}

\item{...}{Additional arguments}

\item{title}{Title for risk calculator (see \code{\link[shiny]{titlePanel}})}

\item{citation}{Citation(s) and author information}

\item{label}{Label for the calculated value}

\item{label_header}{Header label for the result descriptor}

\item{value_header}{Header label for the result value}

\item{format}{Function to format the predicted value for display}

\item{app_name}{App shorthand, like \code{"AppExample"} (https://riskcalc.org/AppExample/)}

\item{labels}{Named character vector specifying labels for any inputs}

\item{levels}{Named list of named character vectors specifying labels for any factor levels}

\item{placeholders}{Named character vector specifying range limits for numeric input variables (of the form \code{c(var="<min>-<max>")}). Invokes calls to \code{\link[shiny]{validate}} enforcing the limits.}
}
\value{
A \code{\link[shiny]{shiny}} application
}
\description{
Construct a formatted \code{\link[shiny]{shiny}}-based risk calculator through different model inputs
}
\examples{
if(interactive()) {
  risk_calculator(glm(1~1))
}
}
