% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/risk_calculator.R
\name{risk_calculator}
\alias{risk_calculator}
\alias{risk_calculator.glm}
\title{Build a risk calculator}
\usage{
risk_calculator(model, ...)

\method{risk_calculator}{glm}(
  model,
  title = "",
  citation = "",
  type = "response",
  label = "Estimated Risk",
  format = function(x) paste0(round(100 * x, 1), "\%"),
  app_name = NULL,
  ...
)
}
\arguments{
\item{model}{A \code{\link[stats]{glm}} object}

\item{...}{Additional arguments}

\item{title}{Title for risk calculator (see \code{\link[shiny]{titlePanel}})}

\item{citation}{Citation(s) and author information}

\item{type}{Response type to be calculated (see the \code{type} argument in \code{\link[stats]{predict.glm}})}

\item{label}{Label for the calculated value}

\item{format}{Function to format the calculated value for display (assumes percentage by default)}

\item{app_name}{App shorthand, like \code{"AppExample"} (https://riskcalc.org/AppExample/)}
}
\value{
A \code{\link[shiny]{shiny}} application
}
\description{
Construct a formatted \code{\link[shiny]{shiny}}-based risk calculator through different model inputs
}
\examples{
if(interactive()) {
  risk_calculator(glm(1~1))
}
}
