% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/riskcalc_fun.R
\name{riskcalc_fun}
\alias{riskcalc_fun}
\title{Title Create a shiny app}
\usage{
riskcalc_fun(x)
}
\arguments{
\item{x}{Unused argument}
}
\value{
A shiny app
}
\description{
Title Create a shiny app
}
\examples{
if(interactive()) {
  riskcalc_fun()
}
}
